# Use buffer to build strings for hashing to noticeably improve the time and memory.