The number of allocs went 15x on the high end of the benchmark.
Also 40% speed up on the high end.

On the more realistic ~25 activators addresses: 50% speedup and 4x reduction in allocations.

/assign @markusthoemmes