@vagababov 2020-04-07
Speedup
```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkSelection/pool-5-subset-1-4        306           314           +2.61%
BenchmarkSelection/pool-5-subset-5-4        310           305           -1.61%
BenchmarkSelection/pool-5-subset-10-4       310           307           -0.97%
BenchmarkSelection/pool-5-subset-15-4       308           309           +0.32%
BenchmarkSelection/pool-5-subset-20-4       309           304           -1.62%
BenchmarkSelection/pool-5-subset-25-4       308           306           -0.65%
BenchmarkSelection/pool-10-subset-1-4       984           998           +1.42%
BenchmarkSelection/pool-10-subset-5-4       978           973           -0.51%
BenchmarkSelection/pool-10-subset-10-4      1222          979           -19.89%
BenchmarkSelection/pool-10-subset-15-4      1088          981           -9.83%
BenchmarkSelection/pool-10-subset-20-4      1046          982           -6.12%
BenchmarkSelection/pool-10-subset-25-4      967           1001          +3.52%
BenchmarkSelection/pool-25-subset-1-4       13519         10134         -25.04%
BenchmarkSelection/pool-25-subset-5-4       13380         9879          -26.17%
BenchmarkSelection/pool-25-subset-10-4      13634         9884          -27.50%
BenchmarkSelection/pool-25-subset-15-4      13366         9979          -25.34%
BenchmarkSelection/pool-25-subset-20-4      13253         9936          -25.03%
BenchmarkSelection/pool-25-subset-25-4      13210         9919          -24.91%
BenchmarkSelection/pool-50-subset-1-4       22805         16307         -28.49%
BenchmarkSelection/pool-50-subset-5-4       22889         16328         -28.66%
BenchmarkSelection/pool-50-subset-10-4      23145         16419         -29.06%
BenchmarkSelection/pool-50-subset-15-4      23306         16107         -30.89%
BenchmarkSelection/pool-50-subset-20-4      22936         16235         -29.22%
BenchmarkSelection/pool-50-subset-25-4      22916         16386         -28.50%
BenchmarkSelection/pool-100-subset-1-4      45189         30134         -33.32%
BenchmarkSelection/pool-100-subset-5-4      44058         30080         -31.73%
BenchmarkSelection/pool-100-subset-10-4     45239         29926         -33.85%
BenchmarkSelection/pool-100-subset-15-4     44969         29879         -33.56%
BenchmarkSelection/pool-100-subset-20-4     44506         30017         -32.56%
BenchmarkSelection/pool-100-subset-25-4     44364         30036         -32.30%
BenchmarkSelection/pool-150-subset-1-4      65451         44912         -31.38%
BenchmarkSelection/pool-150-subset-5-4      65614         45277         -30.99%
BenchmarkSelection/pool-150-subset-10-4     65379         45255         -30.78%
BenchmarkSelection/pool-150-subset-15-4     64335         45324         -29.55%
BenchmarkSelection/pool-150-subset-20-4     64553         44737         -30.70%
BenchmarkSelection/pool-150-subset-25-4     68237         45339         -33.56%
BenchmarkSelection/pool-200-subset-1-4      84962         60848         -28.38%
BenchmarkSelection/pool-200-subset-5-4      85592         60198         -29.67%
BenchmarkSelection/pool-200-subset-10-4     86326         61201         -29.10%
BenchmarkSelection/pool-200-subset-15-4     84658         59864         -29.29%
BenchmarkSelection/pool-200-subset-20-4     91695         60008         -34.56%
BenchmarkSelection/pool-200-subset-25-4     85912         61661         -28.23%
```