@vagababov 2020-04-07
Allocs
```
benchmark                                   old allocs     new allocs     delta
BenchmarkSelection/pool-5-subset-1-4        2              2              +0.00%
BenchmarkSelection/pool-5-subset-5-4        2              2              +0.00%
BenchmarkSelection/pool-5-subset-10-4       2              2              +0.00%
BenchmarkSelection/pool-5-subset-15-4       2              2              +0.00%
BenchmarkSelection/pool-5-subset-20-4       2              2              +0.00%
BenchmarkSelection/pool-5-subset-25-4       2              2              +0.00%
BenchmarkSelection/pool-10-subset-1-4       3              3              +0.00%
BenchmarkSelection/pool-10-subset-5-4       3              3              +0.00%
BenchmarkSelection/pool-10-subset-10-4      3              3              +0.00%
BenchmarkSelection/pool-10-subset-15-4      3              3              +0.00%
BenchmarkSelection/pool-10-subset-20-4      3              3              +0.00%
BenchmarkSelection/pool-10-subset-25-4      3              3              +0.00%
BenchmarkSelection/pool-25-subset-1-4       67             15             -77.61%
BenchmarkSelection/pool-25-subset-5-4       67             15             -77.61%
BenchmarkSelection/pool-25-subset-10-4      67             15             -77.61%
BenchmarkSelection/pool-25-subset-15-4      67             15             -77.61%
BenchmarkSelection/pool-25-subset-20-4      67             15             -77.61%
BenchmarkSelection/pool-25-subset-25-4      67             15             -77.61%
BenchmarkSelection/pool-50-subset-1-4       118            16             -86.44%
BenchmarkSelection/pool-50-subset-5-4       118            16             -86.44%
BenchmarkSelection/pool-50-subset-10-4      118            16             -86.44%
BenchmarkSelection/pool-50-subset-15-4      118            16             -86.44%
BenchmarkSelection/pool-50-subset-20-4      118            16             -86.44%
BenchmarkSelection/pool-50-subset-25-4      118            16             -86.44%
BenchmarkSelection/pool-100-subset-1-4      233            17             -92.70%
BenchmarkSelection/pool-100-subset-5-4      230            17             -92.61%
BenchmarkSelection/pool-100-subset-10-4     230            17             -92.61%
BenchmarkSelection/pool-100-subset-15-4     233            17             -92.70%
BenchmarkSelection/pool-100-subset-20-4     230            17             -92.61%
BenchmarkSelection/pool-100-subset-25-4     230            17             -92.61%
BenchmarkSelection/pool-150-subset-1-4      333            18             -94.59%
BenchmarkSelection/pool-150-subset-5-4      333            18             -94.59%
BenchmarkSelection/pool-150-subset-10-4     336            18             -94.64%
BenchmarkSelection/pool-150-subset-15-4     333            18             -94.59%
BenchmarkSelection/pool-150-subset-20-4     336            18             -94.64%
BenchmarkSelection/pool-150-subset-25-4     333            20             -93.99%
BenchmarkSelection/pool-200-subset-1-4      433            23             -94.69%
BenchmarkSelection/pool-200-subset-5-4      436            24             -94.50%
BenchmarkSelection/pool-200-subset-10-4     436            23             -94.72%
BenchmarkSelection/pool-200-subset-15-4     433            22             -94.92%
BenchmarkSelection/pool-200-subset-20-4     433            24             -94.46%
BenchmarkSelection/pool-200-subset-25-4     436            22             -94.95%
```